REF: move genmod._prediction to base._prediction_inference
