+++ statsmodels/base/_prediction_inference.py
