+++ statsmodels/genmod/generalized_linear_model.py
@@ -27,7 +27,8 @@ from numpy.linalg.linalg import LinAlgError
 import statsmodels.base.model as base
 import statsmodels.base.wrapper as wrap
 
-from statsmodels.genmod._prediction import PredictionResults
+from statsmodels.base import _prediction_inference as pred
+from statsmodels.base._prediction_inference import PredictionResults
 import statsmodels.base._parameter_inference as pinfer
 
 from statsmodels.graphics._regressionplots_doc import (
@@ -50,7 +51,7 @@ from statsmodels.tools.sm_exceptions import (
 from statsmodels.tools.validation import float_like
 
 # need import in module instead of lazily to copy `__doc__`
-from . import _prediction as pred, families
+from . import families
 
 __all__ = ['GLM', 'PredictionResults']
